Progress on #578 

Here are timings of conv2d bench with synchronizes added in:

| branch | fwd | bwd |
| --- | --- | --- |
| main | ~5 ms | ~21.4ms |
| this | ~4.5ms | ~19.75ms |